DNS_With_Django13/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.8
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-extensions==3.1.3
django-filter==21.1
django-guardian==2.4.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
future==0.18.2
idna==3.2
Markdown==3.3.4
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==21.0.0
python-decouple==3.5
python-whois==0.7.3
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.61.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
